Solve username case-insensitiveness issue
The username is always unique and is always strored as lowercases.
It means the username is case-insensitive. However, retrieving users with
the username with non-lowercase letters did not work. Fix the get_user() method
of the memory_repository so that retrieving user by username is case-insensitive
and without trailing spaces.